+++ cif.php
@@ -8,6 +8,14 @@ Parameters:
 
 parse_str($_SERVER["QUERY_STRING"]);
 header("Content-Type: text");
-echo file_get_contents("http://www.crystallography.net/".$codid.".cif");
-?>
+$cif = file_get_contents("http://www.crystallography.net/".$codid.".cif");
 
+if(strpos($cif, "? ? ? ?") == false)
+{
+    echo $cif;
+}
+else
+{
+    header("HTTP/1.0 404 Not Found");
+}
+?>