+++ src/js/Spectroscopy.js
@@ -64,6 +64,7 @@ var Spectroscopy = {
         //update available spectra
         function noSpectra()
         {
+            $("#spectrum-select").empty();
             $("#spectrum-select").append('<option value="default" disabled selected style="display:none;">Choose a spectrum</option>');
             $("#spectrum-select").append('<option value="nmrdb">H1-NMR prediction</option>');
             $("#spectrum").removeClass("loading");